Optimize bitboard and move generation
Add magic bitboards
Reduce perft(6) from over 10s to about 6s

+++ src/move/move.h
@@ -23,7 +23,7 @@ struct Move {
 	uint8_t movedPiece;
 	uint8_t capturedPiece;
 
-	uint64_t newEnPassant;
+	uint8_t newEnPassant;
 
 	Move(){}
